feat: create a CSS file dedicated to site branding
This file is intended to be the single point of configuration
for each individual website. Over time the components will be
refactored to only use colors from this file, allowing the
entire website to be rebranded without running Sass.

BREAKING CHANGE: any sites that were using the older variable
names starting with --implementation will need to be updated
to use --brand. A basic find-and-replace should be sufficient.