fix: adjust appearance of active trail on mobile/desktop
The active trail now appears in both mobile/desktop as
a lighter color for the ancestor items, but the opaque
color for the active menu item only. Additionally, the
active trail treatment is removed from menu items that
have an active-trail parent, but are not themselves
part of the active trail.

Refs: CD-461