fix(a11y): stop announcing redundant SVG data
The SVG is wrapped with an element to hide it from screen
readers because we already have text in the footer to handle
the announcement of the logo. Nevertheless, the SVG's text
tags (title/desc) were also updated to contain genuine info
so that if it is copied out of the website or repo, it will
be announced properly in other contexts.

Refs: CD-476